move hashProduct from utils to mutations
