+++ imports/plugins/core/catalog/server/no-meteor/mutations/hashProduct.test.js
